[Wasm] Inline the `linkIr` function's only usage
Clients (e.g. test facades) should prefer using
`WasmIrLinkingPipelinePhase` instead of calling compiler internals
directly.
